# Pool gzip and lz4 readers and writers