Benchmarks indicate that gzip gets ~20-100% throughput increase
while lz4 goes up ~600-1000%.