@stevevls 2019-02-15
Disclaimer: these numbers were produced on my laptop.

**Gzip**
Before:
```BenchmarkCompression/GZIP1024-8         	   10000	    169127 ns/op	   4.81 MB/s
BenchmarkCompression/GZIP4096-8         	   10000	    219445 ns/op	  14.13 MB/s
BenchmarkCompression/GZIP8192-8         	    5000	    330643 ns/op	  18.65 MB/s
BenchmarkCompression/GZIP16384-8        	    2000	    572457 ns/op	  21.47 MB/s
```

After:
```
BenchmarkCompression/GZIP1024-8         	   20000	     68736 ns/op	  11.80 MB/s
BenchmarkCompression/GZIP4096-8         	   10000	    148396 ns/op	  20.92 MB/s
BenchmarkCompression/GZIP8192-8         	    5000	    247380 ns/op	  24.91 MB/s
BenchmarkCompression/GZIP16384-8        	    3000	    471525 ns/op	  26.05 MB/s
```

**LZ4**
Before:
```
BenchmarkCompression/LZ41024-8          	    1000	   1532284 ns/op	   0.68 MB/s
BenchmarkCompression/LZ44096-8          	    1000	   1748511 ns/op	   2.35 MB/s
BenchmarkCompression/LZ48192-8          	    1000	   1564558 ns/op	   5.25 MB/s
BenchmarkCompression/LZ416384-8         	    1000	   1604422 ns/op	  10.22 MB/s
```

After:
```
BenchmarkCompression/LZ41024-8          	  100000	     15141 ns/op	  68.36 MB/s
BenchmarkCompression/LZ44096-8          	   30000	     52543 ns/op	  78.16 MB/s
BenchmarkCompression/LZ48192-8          	   20000	    116430 ns/op	  70.45 MB/s
BenchmarkCompression/LZ416384-8         	   10000	    254018 ns/op	  64.54 MB/s
```